add payload server test (checks that server returns status 200)
